Make refactoring
Change a directory of source files
